Consolidating utils into core package and updating references
