+++ packages/core/index.ts
@@ -1,1 +1,2 @@
 export * from './src/interfaces'
+export * from './src/app'
