+++ packages/core/src/app/common/utils/constants.ts
@@ -1,5 +1,5 @@
 /* Internal Imports */
-import { getNullString } from './utils'
+import { getNullString } from './misc'
 
 export const NULL_ADDRESS = getNullString(32)
 export const NULL_HASH = getNullString(64)
