+++ packages/core/src/app/common/utils/index.ts
@@ -1,5 +1,8 @@
 export * from './codec'
+export * from './constants'
 export * from './debug-logger'
 export * from './misc'
 export * from './range-store'
+export * from './state-object'
+export * from './sum-tree'
 export * from './type-guards'
