+++ packages/core/src/app/common/utils/misc.ts
@@ -94,6 +94,104 @@ export const bnToUint256 = (bn: BigNum): Buffer => {
  * @returns the transacted range's end as a Uint256 Buffer.
  */
 export const getTransactionRangeEnd = (transaction: Transaction): Buffer => {
-  const end = transaction.stateUpdate.id.end
+  const end = transaction.range.end
   return bnToUint256(end)
+}
+
+/**
+ * Sleeps for a number of milliseconds.
+ * @param ms Number of ms to sleep.
+ * @returns a promise that resolves after the number of ms.
+ */
+export const sleep = (ms: number): Promise<void> => {
+  return new Promise((resolve) => {
+    setTimeout(resolve, ms)
+  })
+}
+
+/**
+ * Removes "0x" from start of a string if it exists.
+ * @param str String to modify.
+ * @returns the string without "0x".
+ */
+export const remove0x = (str: string): string => {
+  return str.startsWith('0x') ? str.slice(2) : str
+}
+
+/**
+ * Adds "0x" to the start of a string if necessary.
+ * @param str String to modify.
+ * @returns the string with "0x".
+ */
+export const add0x = (str: string): string => {
+  return str.startsWith('0x') ? str : '0x' + str
+}
+
+/**
+ * Checks if something is an Object
+ * @param obj Thing that might be an Object.
+ * @returns `true` if the thing is a Object, `false` otherwise.
+ */
+export const isObject = (obj: any): boolean => {
+  return typeof obj === 'object' && obj !== null
+}
+
+/**
+ * Creates a hex string with a certain number of zeroes.
+ * @param n Number of zeroes.
+ * @returns the hex string.
+ */
+export const getNullString = (n: number): string => {
+  return '0x' + '0'.repeat(n)
+}
+
+/**
+ * Reverses a string in place.
+ * @param str String to reverse.
+ * @returns the reversed string.
+ */
+export const reverse = (str: string): string => {
+  return Array.from(str)
+    .reverse()
+    .join('')
+}
+
+/**
+ * Converts a buffer to a hex string.
+ * @param buf the buffer to be converted.
+ * @returns the buffer as a string.
+ */
+export const bufToHexString = (buf: Buffer): string => {
+  return '0x' + buf.toString('hex')
+}
+
+/**
+ * Converts a big number to a hex string.
+ * @param bn the big number to be converted.
+ * @returns the big number as a string.
+ */
+export const bnToHexString = (bn: BigNum): string => {
+  return '0x' + bn.toString('hex')
+}
+
+/**
+ * Converts either a big number or buffer to hex string
+ * @param value the big number or buffer to be converted
+ * @returns the value as a string.
+ */
+export const hexStringify = (value: BigNum | Buffer): string => {
+  if (value instanceof BigNum) {
+    return bnToHexString(value)
+  } else {
+    return bufToHexString(value)
+  }
+}
+
+/**
+ * Converts a hex string to a buffer
+ * @param hexString the hex string to be converted
+ * @returns the hexString as a buffer.
+ */
+export const hexStrToBuf = (hexString: string): Buffer => {
+  return Buffer.from(hexString.slice(2), 'hex')
 }
