+++ packages/core/src/app/common/utils/range-store.ts
@@ -3,15 +3,7 @@ import BigNum = require('bn.js')
 
 /* Internal Imports */
 import { bnMax, bnMin } from './misc'
-
-export interface Range {
-  start: BigNum
-  end: BigNum
-}
-
-export interface BlockRange extends Range {
-  block: BigNum
-}
+import { Range, BlockRange } from '../../../interfaces/common/utils'
 
 /**
  * RangeStore makes it easy to store ranges.
