+++ packages/core/src/interfaces/common/utils/crypto.interface.ts
@@ -1,0 +1,5 @@
+export interface EcdsaSignature {
+  v: string
+  r: string
+  s: string
+}
