+++ packages/core/src/interfaces/common/utils/index.ts
@@ -1,3 +1,4 @@
+export * from './crypto.interface'
 export * from './logger.interface'
 export * from './state.interface'
 export * from './type.interface'
