+++ packages/core/src/interfaces/common/utils/type.interface.ts
@@ -1,1 +1,5 @@
 export type Type<T> = new (...args: any[]) => T
+
+export interface AbiEncodable {
+  encoded: string
+}
