+++ packages/core/test/setup.ts
@@ -1,7 +1,3 @@
 import chai = require('chai')
-import chaiAsPromised = require('chai-as-promised')
 
-const should = chai.should()
-chai.use(chaiAsPromised)
-
-export { should }
+export const should = chai.should()
