+++ packages/predicates/package.json
@@ -31,7 +31,7 @@
   },
   "devDependencies": {
     "@pigi/prettier-config": "^0.0.2-alpha.2",
-    "@pigi/utils": "^0.0.1-alpha.2",
+    "@pigi/core": "^0.0.1-alpha.1",
     "@types/chai": "^4.1.7",
     "@types/debug": "^4.1.2",
     "@types/mocha": "^5.2.6",
