--- packages/utils/package.json
@@ -1,52 +1,0 @@
-{
-  "name": "@pigi/utils",
-  "version": "0.0.1-alpha.10",
-  "description": "Plasma Group core utils",
-  "main": "build/index.js",
-  "files": [
-    "build/**/.js"
-  ],
-  "scripts": {
-    "test": "mocha --require ts-node/register 'test/**/*.spec.ts'",
-    "lint": "tslint --format stylish --project .",
-    "fix": "prettier --config ./.prettierrc.js --write 'index.ts' '{src,test}/**/*.ts'",
-    "build": "tsc -p .",
-    "clean": "rimraf build/"
-  },
-  "keywords": [
-    "plasma",
-    "group",
-    "ethereum",
-    "utils"
-  ],
-  "author": "Plasma Group",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/plasma-group/pigi.git"
-  },
-  "dependencies": {
-    "@types/bn.js": "^4.11.4",
-    "bn.js": "^4.11.8",
-    "ethers": "^4.0.27",
-    "ganache-cli": "^6.4.2",
-    "web3": "^1.0.0-beta.52"
-  },
-  "devDependencies": {
-    "@pigi/prettier-config": "^0.0.2-alpha.2",
-    "@types/chai": "^4.1.7",
-    "@types/debug": "^4.1.2",
-    "@types/mocha": "^5.2.6",
-    "chai": "^4.2.0",
-    "chai-as-promised": "^7.1.1",
-    "mocha": "^6.1.2",
-    "prettier": "^1.16.4",
-    "ts-node": "^8.0.3",
-    "tslint": "^5.15.0",
-    "typescript": "^3.4.2",
-    "rimraf": "^2.6.3"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
