--- packages/utils/src/data-types/index.ts
@@ -1,1 +1,0 @@
-export * from './state-object'
