--- packages/utils/src/index.ts
@@ -1,6 +1,0 @@
-export * from './utils'
-export * from './constants'
-export * from './eth'
-export * from './sum-tree'
-export * from './interfaces'
-export * from './data-types'
