--- packages/utils/test/setup.ts
@@ -1,3 +1,0 @@
-import chai = require('chai')
-
-export const should = chai.should()
