--- packages/utils/tslint.json
@@ -1,6 +1,0 @@
-{
-  "extends": ["./../../tslint.json"],
-  "rules": {
-    "prettier": [true, "./.prettierrc.js"]
-  }
-}
